Made website more responsive, Updated menu
Limited minimum website width to 200px.

Updated the ThreeJs animation to be more fun by adding a trigger to pan
the effect every couple of seconds.

The trigger time is random, but limited to a maximum of 20 seconds,
during which time the effect stops moving as it pans across the screen.