Adding more modular backend
Adding the concept of private rooms to the JS backend. It's actually a
bit harder this far in, but it's not impossible.

As it stands, some core code is commented out, and the main handler is
only a skeleton as of now, but it will get fleshed out really soon.