Added code to generate stages
Got some code off the interwebs which generates mazes, planning on
storing the mazes on the server and allowing the client to request the
maze via ajax ... then we composition the maze on client side with the
correct resources.

Might need to get a tile set from the interwebs to make it a bit more
neat.