Fixed view port calculations
Fixed the point-in-viewport calculation to take into account the width
and height of the object at that point.

Made the World Transform class a bit more safe by adding a variable to
hold the number of fields it has, instead of repeating a number all over
the code, just make reference to the variable.

Added utility functions to the HitRect class to simplify the idea of
expanding and contracting the view rectangle.