Adding collision detection.
Adding collision detection to the code. The first step is to detect if
the player hit any of the enemies. We speed this up by only checking
against enemies which are in the current viewport, this could be sped up
further but not today.

TODO: Debug the collision detection.
TODO: Add detection for the projectiles hitting the enemies.
TODO: Add explosion animation which can be rendered when the enemy or
the player dies.