Added hit testing for enemies and projectiles
Added some collision detection code for enemies and projectiles.
As of right now the code seems to crash my chrome, with no errors or
warnings, which implies a stack overflow of some kind.

TODO: Check if the problem is local to the new code or simply local to
my machine.