Split JS files and added text to interface
Split up JS classes into their own files - yes, there are more http
requests, but it's easier to work with (I can minify the files when the
game is ready for release).

Added two lines of text to the interface - shield and lives. This gives
the player a textual explanation of the bars which appear on the HUD on
the left of the screen.

TODO: Still need to debug the phantom explosions.
TODO: Add power-ups and levels to the game.
TODO: Think a bit more about creating a nice looking HUD panel which
fills the left of the game area.